Add Game Manager, Input Manager, and testing material
Added in Game State system for handling the overall game. Added in a input manager that subscribes and reads values from the input actions object. There is a player movement script and player object and materials, these can be removed before pushing to Main, these were strictly for testing the input scripts.