FMOD Studio Plugin and Audio Manager
Installed the FMOD plugin and created an Audio Manager script that handles all the audio instances and emitters in the game. Created a Music area (can also be for ambience) to change based on which area the player is in. Added in FMODEvents script that lists out all the audio files the game is currently referencing.